Cambio de useReducer por Redux
