+++ src/reducers/GeofencesReducer.js
@@ -1,4 +1,4 @@
-export const GeofencesReducer = (geofences = [], action) => {
+export const geofencesReducer = (geofences = [], action) => {
   switch (action.type){
     case "add":
       return [...geofences, action.payload]
@@ -19,43 +19,8 @@ export const GeofencesReducer = (geofences = [], action) => {
       return geofences.filter(actual => !action.payload.includes(actual._id))
 
     default:
+      return geofences
   }
   
 }
 
-/*
-{
-              geofences.map((geofence)=>{
-                const g = [];
-                
-                geofence.coordinates.map((coord)=>{
-                  g.push([coord.lat, coord.lng]);
-                  return 0;
-                })
-                console.log("g")
-                console.log(g)
-                return <Polygon positions={g} /> 
-              }) 
-            }
-            <InitGeofences geofences={geofences} />
-
-const InitGeofences = ({geofences}) =>{
-    const positions = []
-    if(load === true){
-      geofences.map((geofence)=>{
-        const g = [];
-        
-        geofence.coordinates.map((coord)=>{
-          g.push([coord.lat, coord.lng]);
-          return 0;
-        })
-        console.log("g")
-        console.log(g)
-        positions.push(g)
-        return 0
-      }) 
-      return positions.map(position => <Polygon positions={position} />)
-    }
-  }
-
-            */
